refactor: rely on color.NoColor only
