+++ autocomplete/complete_test.go
@@ -71,7 +71,7 @@ func getApp() *cli.Command {
 }
 
 func getPotentials(cmd string) ([]string, error) {
-	logger := conslogging.Current(conslogging.NoColor, 0, conslogging.Info, false)
+	logger := conslogging.Current(0, conslogging.Info, false)
 	gitLookup := buildcontext.NewGitLookup(logger, "")
 	resolver := buildcontext.NewResolver(nil, gitLookup, logger, "", "", "", 0, "")
