+++ cmd/earthly/main.go
@@ -9,6 +9,7 @@ import (
 	"os"
 	"os/signal"
 	"strconv"
+	"strings"
 	"syscall"
 	"time"
 
@@ -182,10 +183,10 @@ func run() (code int) {
 		}
 	}
 
-	colorMode := conslogging.ColorModeFromEnv()
-
 	// The color package handles NO_COLOR natively. Only unset it for FORCE_COLOR.
-	color.NoColor = colorMode != conslogging.ForceColor && color.NoColor
+	if isForceColor() {
+		color.NoColor = false
+	}
 
 	padding := conslogging.DefaultPadding
 
@@ -210,10 +211,52 @@ func run() (code int) {
 		}
 	}
 
-	logging := conslogging.Current(colorMode, padding, conslogging.Info, cli.Flags().GithubAnnotations)
+	logging := conslogging.Current(padding, conslogging.Info, cli.Flags().GithubAnnotations)
 
 	cli.SetConsole(logging)
 	earth := app.NewEarthApp(cli, rootApp, buildApp)
 
 	return earth.Run(ctx, lastSignal)
+}
+
+// isForceColor returns true if the FORCE_COLOR environment variable is set to a truthy value.
+// It uses a permissive boolean parser to support common truthy/falsy conventions.
+func isForceColor() bool {
+	forceColor := os.Getenv("FORCE_COLOR")
+	if forceColor == "" {
+		return false
+	}
+
+	v, err := parseBool(forceColor)
+	if err != nil {
+		fmt.Printf("read FORCE_COLOR from env: %q\n",
+			forceColor)
+
+		return false
+	}
+
+	return v
+}
+
+func parseBool(val string) (bool, error) {
+	b, err := strconv.ParseBool(val)
+	if err == nil {
+		return b, nil
+	}
+
+	i, err := strconv.Atoi(val)
+	if err == nil {
+		return i != 0, nil
+	}
+
+	switch strings.ToLower(val) {
+	case "yes", "y", "on":
+		return true, nil
+	case "no", "n", "off":
+		return false, nil
+	}
+
+	return false, fmt.Errorf("invalid boolean value, "+
+		"want (1, t, T, TRUE, true, True, 0, f, F, FALSE, false, False, yes, y, on, no, n, off, or integer): %q",
+		val)
 }
