+++ conslogging/conslogging_test.go
@@ -122,88 +122,3 @@ func Test_prettyPrefix(t *testing.T) {
 		})
 	}
 }
-
-func TestColorModeFromEnv(t *testing.T) {
-	testCases := []struct {
-		name       string
-		forceColor string
-		noColor    string
-		want       ColorMode
-	}{
-		{
-			name:       "neither set",
-			forceColor: "",
-			noColor:    "",
-			want:       AutoColor,
-		},
-		{
-			name:       "FORCE_COLOR set to true",
-			forceColor: "1",
-			noColor:    "",
-			want:       ForceColor,
-		},
-		{
-			name:       "FORCE_COLOR set to false",
-			forceColor: "0",
-			noColor:    "",
-			want:       AutoColor,
-		},
-		{
-			name:       "FORCE_COLOR invalid",
-			forceColor: "invalid",
-			noColor:    "",
-			want:       AutoColor,
-		},
-		{
-			name:       "NO_COLOR set to true",
-			forceColor: "",
-			noColor:    "1",
-			want:       NoColor,
-		},
-		{
-			name:       "NO_COLOR set to false",
-			forceColor: "",
-			noColor:    "0",
-			want:       AutoColor,
-		},
-		{
-			name:       "NO_COLOR invalid",
-			forceColor: "",
-			noColor:    "invalid",
-			want:       AutoColor,
-		},
-		{
-			name:       "FORCE_COLOR takes precedence",
-			forceColor: "1",
-			noColor:    "1",
-			want:       ForceColor,
-		},
-		{
-			name:       "FORCE_COLOR false allows NO_COLOR",
-			forceColor: "0",
-			noColor:    "1",
-			want:       NoColor,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Do not use t.Parallel() because t.Setenv modifies process-wide state.
-			if tc.forceColor != "" {
-				t.Setenv("FORCE_COLOR", tc.forceColor)
-			} else {
-				t.Setenv("FORCE_COLOR", "")
-			}
-
-			if tc.noColor != "" {
-				t.Setenv("NO_COLOR", tc.noColor)
-			} else {
-				t.Setenv("NO_COLOR", "")
-			}
-
-			mode := ColorModeFromEnv()
-
-			assert.Equal(t, tc.want, mode)
-		})
-	}
-}
