+++ inputgraph/hash_test.go
@@ -26,7 +26,7 @@ func TestHashTargetWithDocker(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	cons := conslogging.New(os.Stderr, &sync.Mutex{}, conslogging.NoColor, 0, conslogging.Info, false)
+	cons := conslogging.New(os.Stderr, &sync.Mutex{}, 0, conslogging.Info, false)
 
 	hashOpt := HashOpt{Console: cons, Target: target}
 	hash, _, err := HashTarget(ctx, hashOpt)
@@ -118,7 +118,7 @@ func TestHashTargetWithDockerNoAlias(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	cons := conslogging.New(os.Stderr, &sync.Mutex{}, conslogging.NoColor, 0, conslogging.Info, false)
+	cons := conslogging.New(os.Stderr, &sync.Mutex{}, 0, conslogging.Info, false)
 
 	hashOpt := HashOpt{Console: cons, Target: target}
 	hash, _, err := HashTarget(ctx, hashOpt)
@@ -138,7 +138,7 @@ func TestHashTargetWithDockerRemote(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	cons := conslogging.New(os.Stderr, &sync.Mutex{}, conslogging.NoColor, 0, conslogging.Info, false)
+	cons := conslogging.New(os.Stderr, &sync.Mutex{}, 0, conslogging.Info, false)
 
 	hashOpt := HashOpt{Console: cons, Target: target}
 	hash, _, err := HashTarget(ctx, hashOpt)
@@ -157,7 +157,7 @@ func TestHashTargetNoCache(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	cons := conslogging.New(os.Stderr, &sync.Mutex{}, conslogging.NoColor, 0, conslogging.Info, false)
+	cons := conslogging.New(os.Stderr, &sync.Mutex{}, 0, conslogging.Info, false)
 
 	hashOpt := HashOpt{Console: cons, Target: target}
 	hash, stats, err := HashTarget(ctx, hashOpt)
@@ -180,7 +180,7 @@ func TestHashTargetCache(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	cons := conslogging.New(os.Stderr, &sync.Mutex{}, conslogging.NoColor, 0, conslogging.Info, false)
+	cons := conslogging.New(os.Stderr, &sync.Mutex{}, 0, conslogging.Info, false)
 
 	hashOpt := HashOpt{Console: cons, Target: target}
 	hash, stats, err := HashTarget(ctx, hashOpt)
