+++ regproxy/controller_test.go
@@ -12,7 +12,7 @@ func TestNewController(t *testing.T) {
 	t.Parallel()
 
 	// A simple regression test that ensures the values are passed correctly.
-	cons := conslog.Current(conslog.NoColor, 0, conslog.Info, false)
+	cons := conslog.Current(0, conslog.Info, false)
 	c := NewController(nil, nil, true, "proxy-image", time.Second, cons)
 	r := require.New(t)
 	r.Equal("proxy-image", c.darwinProxyImage)
