+++ util/containerutil/frontend_integration_test.go
@@ -882,7 +882,7 @@ func spawnTestVolumes(ctx context.Context, feBinary string, names ...string) (fu
 func testLogger() conslogging.ConsoleLogger {
 	var logs strings.Builder
 
-	logger := conslogging.Current(conslogging.NoColor, conslogging.DefaultPadding, conslogging.Info, false)
+	logger := conslogging.Current(conslogging.DefaultPadding, conslogging.Info, false)
 
 	return logger.WithWriter(&logs)
 }
